Move(#42): GoogleOAuthProperties 패키지 변경
