+++ client/google-oauth-client/src/main/java/com/wypl/googleoauthclient/config/GoogleOAuthProperties.java
@@ -1,4 +1,4 @@
-package com.wypl.googleoauthclient;
+package com.wypl.googleoauthclient.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
